week 12/ToDo/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.0
django-cors-headers==2.5.3
django-filter==2.1.0
django-polymorphic==2.0.3
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
Pillow==7.0.0
psycopg2==2.8.2
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.6.8
pytz==2019.3
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
twilio==6.26.3
typed-ast==1.3.4
urllib3==1.24.3
virtualenv==16.4.3
wrapt==1.11.1
